Version 2
- Remove our own SD card provider because the system comes with one!
- Respect providers that can't delete
- Remove un-needed permissions
- Target API 23

+++ mobile/src/main/java/com/volkhart/localstorage/DeleteActivity.java
@@ -2,8 +2,10 @@ package com.volkhart.localstorage;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.database.Cursor;
 import android.os.Bundle;
 import android.provider.DocumentsContract;
+import android.util.Log;
 import android.widget.Toast;
 
 public class DeleteActivity extends Activity {
@@ -23,8 +25,22 @@ public class DeleteActivity extends Activity {
         switch (requestCode) {
             case RC_DELETE_DOCUMENT:
                 if (resultCode == RESULT_OK) {
-                    DocumentsContract.deleteDocument(getContentResolver(), data.getData());
-                    Toast.makeText(this, R.string.deleted, Toast.LENGTH_SHORT).show();
+                    Cursor cursor = getContentResolver().query(data.getData(), new String[]{DocumentsContract.Document.COLUMN_FLAGS}, null, null, null);
+
+                    if (cursor == null || !cursor.moveToFirst()) {
+                        Toast.makeText(this, R.string.delete_failed, Toast.LENGTH_SHORT).show();
+                        finish();
+                        break;
+                    }
+
+                    int flags = cursor.getInt(0);
+                    if ((DocumentsContract.Document.FLAG_SUPPORTS_DELETE & flags) == DocumentsContract.Document.FLAG_SUPPORTS_DELETE) {
+                        DocumentsContract.deleteDocument(getContentResolver(), data.getData());
+                        Toast.makeText(this, R.string.deleted, Toast.LENGTH_SHORT).show();
+                    } else {
+                        Toast.makeText(this, R.string.delete_failed, Toast.LENGTH_SHORT).show();
+                    }
+                    cursor.close();
                 }
                 finish();
                 break;
